cleanup trace module
