+++ codec/common/src/utils.cpp
@@ -29,80 +29,28 @@
  *     POSSIBILITY OF SUCH DAMAGE.
  *
  *
- * \brief	Tool kits for decoder
- *		( malloc, realloc, free, log output and PSNR calculation and so on )
+ * \file	utils.c
+ *
+ * \brief	common tool/function utilization
  *
  * \date	03/10/2009 Created
  *
  *************************************************************************************
  */
-#ifndef WELS_UTILS_H__
-#define WELS_UTILS_H__
+#include "utils.h"
+#include "crt_util_safe_x.h"	// Safe CRT routines like utils for cross platforms
 
-#include <stdarg.h>
-#include "typedefs.h"
+float WelsCalcPsnr (const void* kpTarPic,
+                    const int32_t kiTarStride,
+                    const void* kpRefPic,
+                    const int32_t kiRefStride,
+                    const int32_t kiWidth,
+                    const int32_t kiHeight);
 
-namespace WelsSVCEnc {
 
-
-/*
- *	Log output routines
- */
-
-typedef int32_t	iWelsLogLevel;
-enum {
-  WELS_LOG_QUIET		= 0x00,		// Quiet mode
-  WELS_LOG_ERROR		= 1 << 0,	// Error log iLevel
-  WELS_LOG_WARNING	= 1 << 1,	// Warning log iLevel
-  WELS_LOG_INFO		= 1 << 2,	// Information log iLevel
-  WELS_LOG_DEBUG		= 1 << 3,	// Debug log iLevel
-  WELS_LOG_RESV		= 1 << 4,	// Resversed log iLevel
-  WELS_LOG_LEVEL_COUNT = 5,
-  WELS_LOG_DEFAULT	= WELS_LOG_ERROR | WELS_LOG_WARNING | WELS_LOG_INFO | WELS_LOG_DEBUG	// Default log iLevel in Wels codec
-};
-
-/*
- *	Function pointer declaration for various tool sets
- */
-// wels log output
-typedef void (*PWelsLogCallbackFunc) (void* pCtx, const int32_t iLevel, const char* kpFmt, va_list argv);
-
-// wels psnr calc
-typedef float (*PWelsPsnrFunc) (const void* kpTarPic,
-                                const int32_t kiTarStride,
-                                const void* kpRefPic,
-                                const int32_t kiRefStride,
-                                const int32_t kiWidth,
-                                const int32_t kiHeight);
-
-extern PWelsLogCallbackFunc	wlog;
-
-#ifdef __GNUC__
-extern void WelsLog (void* pCtx, int32_t iLevel, const char* kpFmt, ...) __attribute__ ((__format__ (__printf__, 3,
-    4)));
-#else
-extern void WelsLog (void* pCtx, int32_t iLevel, const char* kpFmt, ...);
-#endif
-
-extern const char* g_sWelsLogTags[];
-
-/*!
- *************************************************************************************
- * \brief	System trace log output in Wels
- *
- * \param	pCtx	instance pointer
- * \param	kiLevel	log iLevel ( WELS_LOG_QUIET, ERROR, WARNING, INFO, DEBUG )
- * \param	kpFmtStr	formated string to mount
- * \param 	argv	pData string argument
- *
- * \return	NONE
- *
- * \note	N/A
- *************************************************************************************
- */
-void WelsLogDefault (void* pCtx, const int32_t kiLevel, const char* kpFmtStr, va_list argv);
-void WelsLogNil (void* pCtx, const int32_t kiLevel, const char* kpFmtStr, va_list argv);
-
+iWelsLogLevel		g_iLevelLog	= WELS_LOG_DEFAULT;	// default log iLevel
+int32_t			g_iSizeLogBuf	= 1024;			// pBuffer size for each log output
+PWelsLogCallbackFunc	wlog;
 
 /*!
  *************************************************************************************
@@ -115,7 +63,22 @@ void WelsLogNil (void* pCtx, const int32_t kiLevel, const char* kpFmtStr, va_lis
  * \note	can be able to control log iLevel dynamically
  *************************************************************************************
  */
-void WelsSetLogLevel (const int32_t kiLevel);
+void WelsSetLogLevel (const int32_t kiLevel) {
+  iWelsLogLevel iVal = 0;
+  if (kiLevel & WELS_LOG_ERROR) {
+    iVal |= WELS_LOG_ERROR;
+  }
+  if (kiLevel & WELS_LOG_WARNING) {
+    iVal |= WELS_LOG_WARNING;
+  }
+  if (kiLevel & WELS_LOG_INFO) {
+    iVal |= WELS_LOG_INFO;
+  }
+  if (kiLevel & WELS_LOG_DEBUG) {
+    iVal |= WELS_LOG_DEBUG;
+  }
+  g_iLevelLog	= iVal;
+}
 
 /*!
  *************************************************************************************
@@ -128,7 +91,9 @@ void WelsSetLogLevel (const int32_t kiLevel);
  * \note	can be able to get log iLevel of internal codec applicable
  *************************************************************************************
  */
-int32_t WelsGetLogLevel (void);
+int32_t WelsGetLogLevel (void) {
+  return g_iLevelLog;
+}
 
 /*!
  *************************************************************************************
@@ -141,21 +106,25 @@ int32_t WelsGetLogLevel (void);
  * \note	N/A
  *************************************************************************************
  */
-void WelsSetLogCallback (PWelsLogCallbackFunc _log);
+void WelsSetLogCallback (PWelsLogCallbackFunc _log) {
+  wlog	= _log;
+}
 
-/*!
-*************************************************************************************
-* \brief	reopen log file when finish setting current path
-*
-* \param	pCtx		context pCtx
-* \param	pCurPath	current path string
-*
-* \return	NONE
-*
-* \note	N/A
-*************************************************************************************
-*/
-void WelsReopenTraceFile (void* pCtx, char* pCurPath);
+void WelsLogCall (void* pCtx, int32_t iLevel, const char* kpFmt, va_list vl) {
+  wlog (pCtx, iLevel, kpFmt, vl);
+}
+
+void WelsLog (void* pCtx, int32_t iLevel, const char* kpFmt, ...) {
+  va_list vl;
+  va_start (vl, kpFmt);
+  WelsLogCall (pCtx, iLevel, kpFmt, vl);
+  va_end (vl);
+}
+
+#ifndef CALC_PSNR
+#define CONST_FACTOR_PSNR	(10.0 / log(10.0))	// for good computation
+#define CALC_PSNR(w, h, s)	((float)(CONST_FACTOR_PSNR * log( 65025.0 * w * h / iSqe )))
+#endif//CALC_PSNR
 
 /*
  *	PSNR calculation routines
@@ -164,12 +133,12 @@ void WelsReopenTraceFile (void* pCtx, char* pCurPath);
  *************************************************************************************
  * \brief	PSNR calculation utilization in Wels
  *
- * \param	kpTarPic		target picture to be calculated in Picture pData format
- * \param	kiTarStride	stride of target picture pData pBuffer
- * \param 	kpRefPic		base referencing picture samples
- * \param	kiRefStride	stride of reference picture pData pBuffer
- * \param	kiWidth		picture iWidth in pixel
- * \param	kiHeight		picture iHeight in pixel
+ * \param	pTarPic		target picture to be calculated in Picture pData format
+ * \param	iTarStride	stride of target picture pData pBuffer
+ * \param 	pRefPic		base referencing picture samples
+ * \param	iRefStride	stride of reference picture pData pBuffer
+ * \param	iWidth		picture iWidth in pixel
+ * \param	iHeight		picture iHeight in pixel
  *
  * \return	actual PSNR result;
  *
@@ -181,7 +150,24 @@ float WelsCalcPsnr (const void* kpTarPic,
                     const void* kpRefPic,
                     const int32_t kiRefStride,
                     const int32_t kiWidth,
-                    const int32_t kiHeight);
+                    const int32_t kiHeight) {
+  int64_t	iSqe = 0;
+  int32_t x, y;
+  uint8_t* pTar = (uint8_t*)kpTarPic;
+  uint8_t* pRef = (uint8_t*)kpRefPic;
 
+  if (NULL == pTar || NULL == pRef)
+    return (-1.0f);
+
+  for (y = 0; y < kiHeight; ++ y) {	// OPTable !!
+    for (x = 0; x < kiWidth; ++ x) {
+      const int32_t kiT = pTar[y * kiTarStride + x] - pRef[y * kiRefStride + x];
+      iSqe	+= kiT * kiT;
+    }
+  }
+  if (0 == iSqe) {
+    return (99.99f);
+  }
+  return CALC_PSNR (kiWidth, kiHeight, iSqe);
 }
-#endif//WELS_UTILS_H__
+
