+++ codec/decoder/plus/src/welsDecoderExt.cpp
@@ -86,7 +86,7 @@ namespace WelsDec {
 ***************************************************************************/
 CWelsDecoder::CWelsDecoder (void)
   :	m_pDecContext (NULL),
-    m_pTrace (NULL) {
+    m_pWelsTrace (NULL) {
 #ifdef OUTPUT_BIT_STREAM
   char chFileName[1024] = { 0 };  //for .264
   int iBufUsed = 0;
@@ -99,9 +99,20 @@ CWelsDecoder::CWelsDecoder (void)
   int iCurUsedSize;
 #endif//OUTPUT_BIT_STREAM
 
-  m_pTrace = CreateWelsTrace (Wels_Trace_Type);
 
-  IWelsTrace::WelsVTrace (m_pTrace, IWelsTrace::WELS_LOG_INFO, "CWelsDecoder::CWelsDecoder() entry");
+  m_pWelsTrace	= new welsCodecTrace();
+  if (m_pWelsTrace != NULL) {
+    const int32_t iWelsTraceExistingFlag = m_pWelsTrace->WelsTraceModuleIsExist();
+    if (iWelsTraceExistingFlag) {
+      m_pWelsTrace->SetTraceLevel (WELS_LOG_ERROR);
+      WelsSetLogCallback (welsCodecTrace::CODEC_TRACE);
+    }
+  }
+
+  // initialization
+  WelsSetLogLevel (WELS_LOG_ERROR);	// no output, WELS_LOG_QUIET
+
+  WelsLog (NULL, WELS_LOG_INFO, "CWelsDecoder::CWelsDecoder() entry");
 
 #ifdef OUTPUT_BIT_STREAM
   SWelsTime sCurTime;
@@ -156,7 +167,7 @@ CWelsDecoder::CWelsDecoder (void)
 *	return: none
 ***************************************************************************/
 CWelsDecoder::~CWelsDecoder() {
-  IWelsTrace::WelsVTrace (m_pTrace, IWelsTrace::WELS_LOG_INFO, "CWelsDecoder::~CWelsDecoder()");
+  WelsLog (NULL, WELS_LOG_INFO, "CWelsDecoder::~CWelsDecoder()");
 
   UninitDecoder();
 
@@ -171,15 +182,15 @@ CWelsDecoder::~CWelsDecoder() {
   }
 #endif//OUTPUT_BIT_STREAM
 
-  if (NULL != m_pTrace) {
-    delete m_pTrace;
-    m_pTrace = NULL;
+  if (m_pWelsTrace != NULL) {
+    delete m_pWelsTrace;
+    m_pWelsTrace = NULL;
   }
 }
 
 long CWelsDecoder::Initialize (const SDecodingParam* pParam) {
   if (pParam == NULL) {
-    IWelsTrace::WelsVTrace (m_pTrace, IWelsTrace::WELS_LOG_INFO, "CWelsDecoder::Initialize(), invalid input argument.");
+    WelsLog (NULL, WELS_LOG_INFO, "CWelsDecoder::Initialize(), invalid input argument.");
     return cmInitParaError;
   }
 
@@ -201,7 +212,7 @@ void CWelsDecoder::UninitDecoder (void) {
   if (NULL == m_pDecContext)
     return;
 
-  IWelsTrace::WelsVTrace (m_pTrace, IWelsTrace::WELS_LOG_INFO, "into CWelsDecoder::uninit_decoder()..");
+  WelsLog (NULL, WELS_LOG_INFO, "into CWelsDecoder::uninit_decoder()..");
 
   WelsEndDecoder (m_pDecContext);
 
@@ -211,18 +222,19 @@ void CWelsDecoder::UninitDecoder (void) {
     m_pDecContext	= NULL;
   }
 
-  IWelsTrace::WelsVTrace (m_pTrace, IWelsTrace::WELS_LOG_INFO, "left CWelsDecoder::uninit_decoder()..");
+  WelsLog (NULL , WELS_LOG_INFO, "left CWelsDecoder::uninit_decoder()..");
 }
 
 // the return value of this function is not suitable, it need report failure info to upper layer.
 void CWelsDecoder::InitDecoder (void) {
-  IWelsTrace::WelsVTrace (m_pTrace, IWelsTrace::WELS_LOG_INFO, "CWelsDecoder::init_decoder()..");
+
+  WelsLog (NULL, WELS_LOG_INFO, "CWelsDecoder::init_decoder()..");
 
   m_pDecContext	= (PWelsDecoderContext)WelsMalloc (sizeof (SWelsDecoderContext), "m_pDecContext");
 
-  WelsInitDecoder (m_pDecContext, m_pTrace, IWelsTrace::WelsTrace);
+  WelsInitDecoder (m_pDecContext);
 
-  IWelsTrace::WelsVTrace (m_pTrace, IWelsTrace::WELS_LOG_INFO, "CWelsDecoder::init_decoder().. left");
+  WelsLog (NULL, WELS_LOG_INFO, "CWelsDecoder::init_decoder().. left");
 }
 
 /*
@@ -380,8 +392,8 @@ DECODING_STATE CWelsDecoder::DecodeFrame2 (const unsigned char* kpSrc,
       }
     }
 
-    IWelsTrace::WelsVTrace (m_pTrace, IWelsTrace::WELS_LOG_INFO, "decode failed, failure type:%d \n",
-                            m_pDecContext->iErrorCode);
+    WelsLog (NULL, WELS_LOG_INFO, "decode failed, failure type:%d \n",
+             m_pDecContext->iErrorCode);
     return (DECODING_STATE)m_pDecContext->iErrorCode;
   }
