+++ codec/decoder/targets.mk
@@ -19,8 +19,6 @@ DECODER_CPP_SRCS=\
 	$(DECODER_SRCDIR)/core/src/parse_mb_syn_cavlc.cpp\
 	$(DECODER_SRCDIR)/core/src/pic_queue.cpp\
 	$(DECODER_SRCDIR)/core/src/rec_mb.cpp\
-	$(DECODER_SRCDIR)/core/src/utils.cpp\
-	$(DECODER_SRCDIR)/plus/src/welsCodecTrace.cpp\
 	$(DECODER_SRCDIR)/plus/src/welsDecoderExt.cpp\
 
 DECODER_OBJS += $(DECODER_CPP_SRCS:.cpp=.$(OBJ))
