+++ codec/encoder/core/inc/as264_common.h
@@ -148,12 +148,6 @@ $(TargetPath)
 #endif//MEMORY_MONITOR
 #endif//MEMORY_CHECK
 
-//#define ENABLE_TRACE_FILE
-#if defined(_WIN32) && defined(_DEBUG)
-#ifdef ENABLE_TRACE_FILE
-#undef ENABLE_TRACE_FILE
-#endif//ENABLE_TRACE_FILE
-#endif//_WIN32 & _DEBUG
 
 
 #endif // AS264_COMMON_H_
