+++ codec/encoder/targets.mk
@@ -27,9 +27,7 @@ ENCODER_CPP_SRCS=\
 	$(ENCODER_SRCDIR)/core/src/svc_mode_decision.cpp\
 	$(ENCODER_SRCDIR)/core/src/svc_motion_estimate.cpp\
 	$(ENCODER_SRCDIR)/core/src/svc_set_mb_syn_cavlc.cpp\
-	$(ENCODER_SRCDIR)/core/src/utils.cpp\
 	$(ENCODER_SRCDIR)/core/src/wels_preprocess.cpp\
-	$(ENCODER_SRCDIR)/plus/src/welsCodecTrace.cpp\
 	$(ENCODER_SRCDIR)/plus/src/welsEncoderExt.cpp\
 
 ENCODER_OBJS += $(ENCODER_CPP_SRCS:.cpp=.$(OBJ))
